Fixed missing class definition for LinkState in link.py
